requirements.txt
absl-py==2.1.0
filelock==3.18.0
fsspec==2025.3.0
grpcio==1.71.0
Jinja2==3.1.6
joblib==1.4.2
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
protobuf==6.30.1
scikit-learn==1.6.1
scipy==1.15.2
setuptools==76.0.0
six==1.17.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
Werkzeug==3.1.3
